pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "signaldeck"
version = "0.1.0"
description = "SDR scanner, classifier, and decoder platform"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.26",
    "scipy>=1.11",
    "aiosqlite>=0.19",
    "click>=8.1",
    "pyyaml>=6.0",
    "bcrypt>=4.1",
    "fastapi>=0.100",
    "uvicorn>=0.20",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[project.scripts]
signaldeck = "signaldeck.main:cli"

[tool.setuptools.packages.find]
include = ["signaldeck*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "hardware: requires connected SDR hardware",
]
